refraction: force user awareness about converting to PyMVG
There is a fundamental question about how to represent refractive
boundaries. One could argue that they don't belong with the camera
calibration as they are not part of the cameras per se. This is the
case with PyMVG. With flydra, however, we have a "Reconstructor" to
represent a multi-camera system and also potentially a refractive
boundary. So, to export from a Reconstructor to a PyMVG camera system,
we need to drop the representation of the refractive boundary. We
should force the user to know that she is dropping this information
when doing so.